test: move to tests/
